weather-tools/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
google-api-python-client>=2.125.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
